Bug 20844: Fixes problem sending file > 4G on 32 bit systems
Before we were using a size_t for payload size, which worked fine on 64
bit systems, but not on 32 bit systems. So now we are using uint64_t
instead.

+++ src/db_layer/src/jaldb_serialize_record.c
@@ -505,7 +505,7 @@ enum jaldb_status jaldb_deserialize_string(uint8_t **buffer, size_t *size, char*
 }
 
 enum jaldb_status jaldb_deserialize_segment(char on_disk,
-		size_t segment_length,
+		uint64_t segment_length,
 		uint8_t **buffer,
 		size_t *bsize,
 		struct jaldb_segment **segment)
