+++ src/db_layer/src/jaldb_serialize_record.h
@@ -189,7 +189,7 @@ enum jaldb_status jaldb_deserialize_string(uint8_t **buffer, size_t *size, char*
  * @return JALDB_OK on success, or an error.
  */
 enum jaldb_status jaldb_deserialize_segment(char on_disk,
-		size_t segment_length,
+		uint64_t segment_length,
 		uint8_t **buffer,
 		size_t *size,
 		struct jaldb_segment **segment);
